requirements.txt
pyautogui
nfcpy